+++ Code/GraphMol/ChemReactions/testReaction.cpp
@@ -6853,6 +6853,31 @@ void testGithub1269() {
   }
 }
 
+void testGithub1868() {
+  BOOST_LOG(rdInfoLog) << "-------------------------------------" << std::endl;
+  BOOST_LOG(rdInfoLog) << "Testing Github #1868: Atom index out of range error"
+                       << std::endl;
+
+  std::string rdbase = getenv("RDBASE");
+  std::string fName;
+
+  fName = rdbase + "/Code/GraphMol/ChemReactions/testData/v3k.AmideBond.rxn";
+
+  for (int i = 0; i < 1000; ++i) {
+    std::unique_ptr<ChemicalReaction> rxn(RxnFileToChemicalReaction(fName));
+    TEST_ASSERT(rxn);
+    TEST_ASSERT(rxn->getNumReactantTemplates() == 2);
+    TEST_ASSERT(rxn->getNumProductTemplates() == 1);
+
+    for (auto v : rxn->getReactants()) {
+      MolToSmiles(*v.get());
+    }
+    for (auto v : rxn->getProducts()) {
+      MolToSmiles(*v.get());
+    }
+  }
+}
+
 int main() {
   RDLog::InitLogs();
 
@@ -6938,9 +6963,9 @@ int main() {
   testReactionProperties();
   testGithub1950();
   testGithub1869();
-#endif
   testGithub1269();
-
+#endif
+  testGithub1868();
   BOOST_LOG(rdInfoLog)
       << "*******************************************************\n";
   return (0);