+++ Code/GraphMol/new_canon.cpp
@@ -21,6 +21,10 @@ namespace RDKit {
 namespace Canon {
 void CreateSinglePartition(unsigned int nAtoms, int *order, int *count,
                            canon_atom *atoms) {
+  PRECONDITION(order, "bad pointer");
+  PRECONDITION(count, "bad pointer");
+  PRECONDITION(atoms, "bad pointer");
+
   for (unsigned int i = 0; i < nAtoms; i++) {
     atoms[i].index = 0;
     order[i] = i;
@@ -31,6 +35,10 @@ void CreateSinglePartition(unsigned int nAtoms, int *order, int *count,
 
 void ActivatePartitions(unsigned int nAtoms, int *order, int *count,
                         int &activeset, int *next, int *changed) {
+  PRECONDITION(order, "bad pointer");
+  PRECONDITION(count, "bad pointer");
+  PRECONDITION(next, "bad pointer");
+  PRECONDITION(changed, "bad pointer");
   unsigned int i, j;
   activeset = -1;
   for (i = 0; i < nAtoms; i++) next[i] = -2;
@@ -63,10 +71,12 @@ void ActivatePartitions(unsigned int nAtoms, int *order, int *count,
 void compareRingAtomsConcerningNumNeighbors(Canon::canon_atom *atoms,
                                             unsigned int nAtoms,
                                             const ROMol &mol) {
+  PRECONDITION(atoms, "bad pointer");
   RingInfo *ringInfo = mol.getRingInfo();
   for (unsigned idx = 0; idx < nAtoms; ++idx) {
     const Canon::canon_atom &a = atoms[idx];
-    if (!ringInfo->isInitialized() || ringInfo->numAtomRings(a.atom->getIdx()) < 1) {
+    if (!ringInfo->isInitialized() ||
+        ringInfo->numAtomRings(a.atom->getIdx()) < 1) {
       continue;
     }
     std::deque<int> neighbors;
@@ -75,14 +85,18 @@ void compareRingAtomsConcerningNumNeighbors(Canon::canon_atom *atoms,
     atoms[idx].neighborNum.reserve(1000);
     atoms[idx].revistedNeighbors.assign(1000, 0);
     char *visited = (char *)malloc(nAtoms * sizeof(char));
+    CHECK_INVARIANT(visited, "allocation failed");
     memset(visited, 0, nAtoms * sizeof(char));
     unsigned count = 1;
     std::vector<int> nextLevelNbrs;
     char *lastLevelNbrs = (char *)malloc(nAtoms * sizeof(char));
+    CHECK_INVARIANT(lastLevelNbrs, "allocation failed");
     memset(lastLevelNbrs, 0, nAtoms * sizeof(char));
     char *currentLevelNbrs = (char *)malloc(nAtoms * sizeof(char));
+    CHECK_INVARIANT(currentLevelNbrs, "allocation failed");
     memset(currentLevelNbrs, 0, nAtoms * sizeof(char));
     int *revisitedNeighbors = (int *)malloc(nAtoms * sizeof(int));
+    CHECK_INVARIANT(revisitedNeighbors, "allocation failed");
     memset(revisitedNeighbors, 0, nAtoms * sizeof(int));
     while (!neighbors.empty()) {
       unsigned int numLevelNbrs = 0;
@@ -91,7 +105,8 @@ void compareRingAtomsConcerningNumNeighbors(Canon::canon_atom *atoms,
         int nidx = neighbors.front();
         neighbors.pop_front();
         const Canon::canon_atom &atom = atoms[nidx];
-        if (!ringInfo->isInitialized() || ringInfo->numAtomRings(atom.atom->getIdx()) < 1) {
+        if (!ringInfo->isInitialized() ||
+            ringInfo->numAtomRings(atom.atom->getIdx()) < 1) {
           continue;
         }
         lastLevelNbrs[nidx] = 1;
@@ -164,14 +179,19 @@ void rankWithFunctor(T &ftor, bool breakTies, int *order,
                      bool useSpecial = false, bool useChirality = false,
                      const boost::dynamic_bitset<> *atomsInPlay = nullptr,
                      const boost::dynamic_bitset<> *bondsInPlay = nullptr) {
+  PRECONDITION(order, "bad pointer");
   const ROMol &mol = *ftor.dp_mol;
   canon_atom *atoms = ftor.dp_atoms;
   unsigned int nAts = mol.getNumAtoms();
   int *count = (int *)malloc(nAts * sizeof(int));
+  CHECK_INVARIANT(count, "allocation failed");
   int activeset;
   int *next = (int *)malloc(nAts * sizeof(int));
+  CHECK_INVARIANT(next, "allocation failed");
   int *changed = (int *)malloc(nAts * sizeof(int));
+  CHECK_INVARIANT(changed, "allocation failed");
   char *touched = (char *)malloc(nAts * sizeof(char));
+  CHECK_INVARIANT(touched, "allocation failed");
   memset(touched, 0, nAts * sizeof(char));
   memset(changed, 1, nAts * sizeof(int));
   CreateSinglePartition(nAts, order, count, atoms);
@@ -237,7 +257,8 @@ void rankWithFunctor(T &ftor, bool breakTies, int *order,
         symRingAtoms += count[order[i]];
       }
       ringAtoms++;
-      if (ringInfo->isInitialized() && ringInfo->numAtomRings(order[i]) > 1 && count[order[i]] > 1) {
+      if (ringInfo->isInitialized() && ringInfo->numAtomRings(order[i]) > 1 &&
+          count[order[i]] > 1) {
         branchingRingAtom = true;
       }
     }
@@ -283,10 +304,11 @@ void rankWithFunctor(T &ftor, bool breakTies, int *order,
 
 namespace {
 bool hasRingNbr(const ROMol &mol, const Atom *at) {
+  PRECONDITION(at, "bad pointer");
   ROMol::ADJ_ITER beg, end;
   boost::tie(beg, end) = mol.getAtomNeighbors(at);
   while (beg != end) {
-    const Atom* nbr = mol[*beg];
+    const Atom *nbr = mol[*beg];
     ++beg;
     if ((nbr->getChiralTag() == Atom::CHI_TETRAHEDRAL_CW ||
          nbr->getChiralTag() == Atom::CHI_TETRAHEDRAL_CCW) &&
@@ -298,12 +320,14 @@ bool hasRingNbr(const ROMol &mol, const Atom *at) {
 }
 
 void getNbrs(const ROMol &mol, const Atom *at, int *ids) {
+  PRECONDITION(at, "bad pointer");
+  PRECONDITION(ids, "bad pointer");
   ROMol::OEDGE_ITER beg, end;
   boost::tie(beg, end) = mol.getAtomBonds(at);
   unsigned int idx = 0;
 
   while (beg != end) {
-    const Bond* bond = (mol)[*beg];
+    const Bond *bond = (mol)[*beg];
     ++beg;
     unsigned int nbrIdx = bond->getOtherAtomIdx(at->getIdx());
     ids[idx] = nbrIdx;
@@ -313,6 +337,7 @@ void getNbrs(const ROMol &mol, const Atom *at, int *ids) {
 
 bondholder makeBondHolder(const Bond *bond, unsigned int otherIdx,
                           bool includeChirality) {
+  PRECONDITION(bond, "bad pointer");
   Bond::BondStereo stereo = Bond::STEREONONE;
   if (includeChirality) {
     stereo = bond->getStereo();
@@ -326,23 +351,25 @@ bondholder makeBondHolder(const Bond *bond, unsigned int otherIdx,
 }
 void getBonds(const ROMol &mol, const Atom *at, std::vector<bondholder> &nbrs,
               bool includeChirality) {
+  PRECONDITION(at, "bad pointer");
   ROMol::OEDGE_ITER beg, end;
   boost::tie(beg, end) = mol.getAtomBonds(at);
   while (beg != end) {
-    const Bond* bond = (mol)[*beg];
+    const Bond *bond = (mol)[*beg];
     ++beg;
-    nbrs.push_back(makeBondHolder(
-        bond, bond->getOtherAtomIdx(at->getIdx()), includeChirality));
+    nbrs.push_back(makeBondHolder(bond, bond->getOtherAtomIdx(at->getIdx()),
+                                  includeChirality));
   }
   std::sort(nbrs.begin(), nbrs.end(), bondholder::greater);
 }
 
 void getChiralBonds(const ROMol &mol, const Atom *at,
                     std::vector<bondholder> &nbrs) {
+  PRECONDITION(at, "bad pointer");
   ROMol::OEDGE_ITER beg, end;
   boost::tie(beg, end) = mol.getAtomBonds(at);
   while (beg != end) {
-    const Bond* bond = (mol)[*beg];
+    const Bond *bond = (mol)[*beg];
     ++beg;
     unsigned int nbrIdx = bond->getOtherAtomIdx(at->getIdx());
     const Atom *nbr = mol.getAtomWithIdx(nbrIdx);
@@ -509,6 +536,7 @@ void freeCanonAtoms(std::vector<Canon::canon_atom> &atoms) {
 }
 
 void updateAtomNeighborIndex(canon_atom *atoms, std::vector<bondholder> &nbrs) {
+  PRECONDITION(atoms, "bad pointer");
   for (auto &nbr : nbrs) {
     unsigned nbrIdx = nbr.nbrIdx;
     unsigned newSymClass = atoms[nbrIdx].index;
@@ -559,12 +587,11 @@ void updateAtomNeighborNumSwaps(
 }
 
 void rankMolAtoms(const ROMol &mol, std::vector<unsigned int> &res,
-                  bool breakTies, bool includeChirality, bool includeIsotopes) { 
+                  bool breakTies, bool includeChirality, bool includeIsotopes) {
   res.resize(mol.getNumAtoms());
- 
-  if (!mol.getNumAtoms())
-    return;
-  
+
+  if (!mol.getNumAtoms()) return;
+
   std::vector<Canon::canon_atom> atoms(mol.getNumAtoms());
   initCanonAtoms(mol, atoms, includeChirality);
   AtomCompareFunctor ftor(&atoms.front(), mol);
@@ -573,6 +600,7 @@ void rankMolAtoms(const ROMol &mol, std::vector<unsigned int> &res,
   ftor.df_useChiralityRings = includeChirality;
 
   int *order = (int *)malloc(mol.getNumAtoms() * sizeof(int));
+  PRECONDITION(order, "bad pointer");
   rankWithFunctor(ftor, breakTies, order, true, includeChirality);
 
   for (unsigned int i = 0; i < mol.getNumAtoms(); ++i) {
@@ -594,10 +622,9 @@ void rankFragmentAtoms(const ROMol &mol, std::vector<unsigned int> &res,
                "bad atomSymbols size");
 
   res.resize(mol.getNumAtoms());
-  
-  if (!mol.getNumAtoms())
-    return;
-        
+
+  if (!mol.getNumAtoms()) return;
+
   std::vector<Canon::canon_atom> atoms(mol.getNumAtoms());
   initFragmentCanonAtoms(mol, atoms, includeChirality, atomSymbols, atomsInPlay,
                          bondsInPlay);
@@ -607,7 +634,7 @@ void rankFragmentAtoms(const ROMol &mol, std::vector<unsigned int> &res,
   ftor.df_useChirality = includeChirality;
 
   int *order = (int *)malloc(mol.getNumAtoms() * sizeof(int));
-
+  PRECONDITION(order, "bad pointer");
   rankWithFunctor(ftor, breakTies, order, true, includeChirality, &atomsInPlay,
                   &bondsInPlay);
 
@@ -621,14 +648,14 @@ void rankFragmentAtoms(const ROMol &mol, std::vector<unsigned int> &res,
 void chiralRankMolAtoms(const ROMol &mol, std::vector<unsigned int> &res) {
   res.resize(mol.getNumAtoms());
 
-  if(!mol.getNumAtoms())
-    return;
-  
+  if (!mol.getNumAtoms()) return;
+
   std::vector<Canon::canon_atom> atoms(mol.getNumAtoms());
   initChiralCanonAtoms(mol, atoms);
   ChiralAtomCompareFunctor ftor(&atoms.front(), mol);
 
   int *order = (int *)malloc(mol.getNumAtoms() * sizeof(int));
+  PRECONDITION(order, "bad pointer");
   rankWithFunctor(ftor, false, order);
 
   for (unsigned int i = 0; i < mol.getNumAtoms(); ++i) {
@@ -637,5 +664,5 @@ void chiralRankMolAtoms(const ROMol &mol, std::vector<unsigned int> &res) {
   free(order);
   freeCanonAtoms(atoms);
 }  // end of rankMolAtoms()
-}  // end of Canon namespace
-}  // end of RDKit namespace
+}  // namespace Canon
+}  // namespace RDKit