+++ Code/JavaWrappers/gmwrapper/src-test/org/RDKit/BasicMoleculeTests.java
@@ -243,6 +243,16 @@ public class BasicMoleculeTests extends GraphMolTest {
 
 	}
 
+	@Test public void testGetAtomNeighbors() {
+		String smiles="CC(C)C";
+		ROMol mol = RWMol.MolFromSmiles(smiles);
+		assertEquals(mol.getAtomNeighbors(mol.getAtomWithIdx(0)).size(),1);
+		assertEquals(mol.getAtomNeighbors(mol.getAtomWithIdx(1)).size(),3);
+		assertEquals(mol.getAtomNeighbors(mol.getAtomWithIdx(2)).size(),1);
+		assertEquals(mol.getAtomBonds(mol.getAtomWithIdx(0)).size(),1);
+		assertEquals(mol.getAtomBonds(mol.getAtomWithIdx(1)).size(),3);
+		assertEquals(mol.getAtomBonds(mol.getAtomWithIdx(2)).size(),1);
+	}
 
 /*	@Test -- the contents of this test now in UnitTestPickling, testIssue219
 	public void testConformer(){